requirements.txt
jsonschema==3.0.2
tqdm==4.50.2
Flask_DebugToolbar==0.11.0
chartkick==0.5.0
CensusData==1.9
plydata==0.4.3
WTForms_Components==0.10.4
WTForms==2.2.1
